api/requirements.txt
asgiref==3.9.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.4
django-appconf==1.1.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
dpd-static-support==0.0.5
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
narwhals==2.0.1
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
retrying==1.4.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.9.0
zipp==3.23.0
